Update Dl4EO: Using URI + all code in 1 file
