+++ public/DL4EO_Airplane_Detection/meta.json
@@ -59,12 +59,7 @@
             "fused:description": "## Overview\n\nThis UDF runs an object detection model on 50cm imagery. \n\nThe workflow uses the `ultralytics` library to run a `YOLO` model on an RGB tile of a satellite image and returns a GeoDataFrame with bounding boxes for the predicted objects.\n\nYou can find the image in the [Public ArcGIS RGB UDF](https://www.fused.io/workbench/catalog/Arcgis_Rgb-b58b7a3c-3dc0-4e9d-9bdd-7092f3ea9817). For best visualization effects, render the output of that UDF as layer underneath.\n\nThe development of this UDF was led by [Jeff Faudi](https://www.linkedin.com/in/faudi/) @ DL4EO. For a detailed overview of the methods and applications, refer to Jeff's [blog post](https://docs.fused.io/blog/ai-for-object-detection-on-50cm-imagery/).\n\n\n"
           },
           "source": "DL4EO_Airplane_Detection.py",
-          "headers": [
-            {
-              "module_name": "utils",
-              "source_file": "utils.py"
-            }
-          ]
+          "headers": []
         }
       }
     ],
